ui: send request state to component, process data in components
We should not decouple the data from a fetch request and fields
describing the request/response like `inFlight`, `valid`,
`lastUpdated` which are stored in redux, when passing them as
props into the component.

In traditional redux apps we we would be able to select the
request state as a whole unit, getting fields like `isLoading`,
`error` and etc with the data portion of the payload.
Unfortunately, with the our split redux stores in db-console
and CC, we were not able to select directly from the stores.
Instead, we have been following a pattern of creating selectors
for each required part of the request field (e.g. `selectData`,
`selectError`, `selectDataIsValid`) for each cached request,
having to create a copy of each selector in both db-console
and cluster-ui. These selectors are then used to pass each field
independently as a prop. One can see how this boilerplate code can
quickly multiply with the number of new requests being created.

This pattern was probably a mistake -- why do we use selectors
in the first place? To memoize computations when we want to
transform data from the redux store. However, for modern React
we have built-in utilities like `useMemo` that can accomplish
this within the component. For fields like `isLoading`, where
we just read some fields from redux with no computation, we don't
need to cache anything, so we're currently requiring all this
boilerplate with no useful effects. In addition, the processing
of some of the data in each of these selectors (db-console and
cluster-ui) means we have code that gets duplicated. This is hard
to maintain, as that data, when making it into the component, may
even go through more processing. Let's try to keep the processing
in the component itself, containing it to more central place.

We should adopt a pattern of sending the entire request state
for cached data fetches to the components reading them. This
commit introduces the `RequestState<T>>` type which should be
used as the redux field type for all cached data fetches.

This commit shows an example of what can be done with the
statements page and response. Note that for class components,
where `useMemo` is not available, we can just use redux's
selectors/reselect to help memoize when necessary.

Epic: none

Release note: None