sqlstats: buffer statement stats in regular slice instead of sqlstats container
Previously, each stats collector for a sql session  would buffer a
transaciton's stmt stats in a ssmemstorage.Container struct which has
a lot of additional overhead when reading and writing as it manages
synchronization logic. For this per-session writing, this writing
is synchronous so there's no need to use such a container. We should
just buffer the stmts in a simple slice. Essentially, this shouldn't
make anything worse as we're just simplifying the buffer.

Note that this is simply a temporary simplified state. Eventually as
part of #141024 we will move the writing to the shared container to be
async from query execution.

Epic: none

Release note: None