sqlstats: generalize insight.ConcurrentBufferIngester
This commit copies the ConcurrentBufferIngester from
insights into the sqlstats pkg in preparation to have it
consume the sql execution event and handle the the writing
 to both sqlstats and insights systems, instead of just
for insights. At this stage we are preserving how this
ingester consumes events. The ingester is still only used
for insights writing.

Epic: none

Release note: None