insights: insights and sqlstats subsystem should consume same structs
Previously, sqlstats and insights consumed different struct types for
their recording step. These should only be created if we determine
that the execution is an outlier.

SQL stats and insights will now consume the same data types to
process and perform their respective recordings: `sqlstats.RecordedStmtStats`.
and `sqlstats.RecordedTxnStats`. This is supporting work in moving
sql stats recording to to occur in the same async routine as insights.

Epic: none
Release note: None